Extract top-level helpers into src/helpers.js
Move 13 standalone helper functions out of thefade.js into a dedicated
module so the entry file is closer to just classes + hooks. These
functions have no `this` context and no class references, so extraction
is a pure relocation — no behavior change.

Extracted: getRankValue, applySpellFilters, openCompendiumBrowser,
initializeDefaultSkills, createCustomSkill, showCustomSkillDialog,
applyPathSkillModifications, improveCharacterSkill,
handleChooseSkillsOption, showChooseRegularSkillsDialog,
showChooseLoreSkillsDialog, showChoosePerformSkillsDialog,
showChooseCraftSkillsDialog.

`registerItemSheets` stays in thefade.js because it references the
TheFadeItemSheet class defined there.

thefade.js: 8800 → 7934 lines.
